Update pid_tuning_guide_multicopter.md
Some changes to the documentation. The first is a mistake whilst the others are more suggested improvements.

I also did not get the any of the responses in this documentation with my platform but maybe I did something wrong. Default gains for MC provided a very decent response.

Lastly I think there could be a suggestion on "Keep the multi rotor in your hand and increase the thrust to about 50%,
so that the weight is virtually zero." for people that test bigger (and heavier) multicopters? XD