Adding a note to raise awareness for throttle input (#1190)
* Adding a not to raise awareness for throttle input

It took me a very long time to figure out, why I couldn't use my summing mixers in disarmed state. I knew that throttle signals wouldn't work in disarmed state, but I had assumed that a summing mixer that gets multiple control inputs would still work in disarmed state for all signals that are NOT throttle. Since that is not the case and I assume others might have the same issue, I would like to raise awareness for it. I'm not sure, if this would be the right place to do so, but I decided to just give it a try.

* rewrite note

Co-authored-by: Hamish Willee <hamishwillee@gmail.com>