Removed link to inactive webshop
Reason: The linked webshop still seems to operate (based on the website), but the shop itself is not. I ordered through the webshop but nothing ever shipped and emails are not being answered. According to a forum posts ( https://community.shopify.com/c/shopify-discussions/how-to-reach-out-a-non-performing-vendor-how-to-understand-a/m-p/1266655/highlight/true#M252691 ) on the web this is not the first instance of not delivered products.