Remove deprecated information on Iris lockstep (#1761)
* Remove deprecated information on Iris lockstep

The Xacro file doesn't seem to exist anymore, neither does the `models/rotors_model`-path. If you want to turn off lockstep for Iris, you can do that in the SDF file as well. So this additional information is deprecated and confusing.

* Complete the config command

Add `make ` to the config command in order to have the full command there and make it easier for people who are not entirely familiar with PX4 workflow.