flight_controller: add Orange+, edit Orange/Yellow
This adds docs for the CubePilot Cube Orange+, and updates the docs and
links for Orange and Yellow while I'm at it.

I remove the the note regarding Cube not being recommended as it's not
based on the Pixhawk standard since the Cubes are already listed as
"manufacturer supported" rathen than being standard Pixhawks, so that
should be enough.

Signed-off-by: Julian Oes <julian@oes.ch>