esc_calibration: document latest functionality
For the 1.14 release because we already changed the actuator
configuration and users will need to go through their setups I also
corrected the PWM calibration and the default PWM values.

See https://github.com/PX4/PX4-Autopilot/pull/21711
and https://github.com/PX4/PX4-Autopilot/pull/21513

This is the documentation change with it. And needs to be
referenced by the release notes.